+++ lib/requests_job.php
@@ -39,7 +39,15 @@ function dispatch_request($request)
 
             require_once(DIR_OOL.'job_list.php');
 
-            $job_list_data = jobs_list_retrieve($DBH);
+            // retrieving filters data
+            // check and set all passed parameters
+            $filters = array('job_category_level_1'  => post_or_get('a'),    // category
+                             'job_category_level_2'  => post_or_get('i'),    // issue
+                             'job_creation_datetime' => post_or_get('x'),    // creation date
+                             'job_status'            => post_or_get('s')    // status
+                             );
+
+            $job_list_data = jobs_list_retrieve($DBH, $filters);
 
             if ( count($job_list_data) == 0 ) {
                 $retval['success'] = false;