+++ oolib/job_list.php
@@ -1,7 +1,30 @@
 <?php
 
-function jobs_list_retrieve($DBH)
+function jobs_list_retrieve($DBH, $filters)
 {
+    $job_category_level_1 =
+        ( $filters['job_category_level_1'] === false ?
+          ''                                          :
+          '   AND job_category_level_1 = '.$filters['job_category_level_1'].' '
+        );
+    $job_category_level_2  =
+        ( $filters['job_category_level_2'] === false ?
+          ''                                         :
+          '   AND job_category_level_2 = '.$filters['job_category_level_2'].' '
+        );
+    $job_creation_datetime =
+        ( $filters['job_creation_datetime'] === false ?
+          ''                                          :
+          '   AND job_creation_datetime >= '.$filters['job_creation_datetime'].' '
+        );
+    $job_status =
+        ( $filters['job_status'] === false ?
+          ''                               :
+          '   AND job_status = '.$filters['job_status'].' '
+        );
+
+
+
     $query = 'SELECT Job_id, Job_subject, Job_description, Job_start_datetime, Job_priority, '.
              '       Status_name, '.
              '       Category_1_name, Category_2_name, '.
@@ -10,7 +33,13 @@ function jobs_list_retrieve($DBH)
              '                    INNER JOIN '.TBL_CATEGORIES_2.' ON '.TBL_JOBS.'.Job_category_level_2 = '.TBL_CATEGORIES_2.'.Category_2_id '.
              '                    INNER JOIN '.TBL_STATUSES.'     ON '.TBL_JOBS.'.Job_status           = '.TBL_STATUSES.'.Status_id '.
              '                    INNER JOIN '.TBL_USERS.'        ON '.TBL_JOBS.'.Job_assigned_to_User = '.TBL_USERS.'.User_id '.
-             ' WHERE Job_trashed <> 1 ';
+             ' WHERE 1 '.
+             $job_category_level_1.
+             $job_category_level_2.
+             $job_creation_datetime.
+             $job_status.
+             '   AND Job_trashed <> 1 ';
+
 
     $stmt_read = $DBH->prepare($query);
     $stmt_read->execute();