+++ CHANGES.md
@@ -11,7 +11,7 @@ TODOS:
 - now also the job_creation_date select form field have its object in job_search class
 - implemented update_filters_values method in search_filters class (should improve it)
 - added a filter_values property in search_filters class
-- the DATA__load_job_list method in job_list class now have parameters set by search filters status
+- now DATA__load_job_list method in job_list class have parameters set by search filters status sent in the Ajax call
 
 19.05.2013
 ----------